+++ app/src/androidTest/java/com/syabonbubble/profilecode/ExampleInstrumentedTest.kt
@@ -1,0 +1,24 @@
+package com.syabonbubble.profilecode
+
+import androidx.test.InstrumentationRegistry
+import androidx.test.runner.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.junit.Assert.*
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        assertEquals("jp.co.cyberagent.dojo2019", appContext.packageName)
+    }
+}
