+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="jp.co.cyberagent.dojo2019">
+          package="com.syabonbubble.profilecode">
 
     <uses-permission android:name="android.permission.CAMERA"/>
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
@@ -15,12 +15,12 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".Activity.WebActivity">
+        <activity android:name=".activity.WebActivity">
         </activity>
-        <activity android:name=".Activity.QRscannerActivity">
+        <activity android:name=".activity.QRscannerActivity">
         </activity>
         <activity
-                android:name=".Activity.MainActivity">
+                android:name=".activity.MainActivity">
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
