+++ app/src/main/res/layout/activity_qrscanner.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activity.QRscannerActivity">
+        tools:context=".activity.QRscannerActivity">
 
     <com.journeyapps.barcodescanner.CompoundBarcodeView
             android:id="@+id/qr_view"
