+++ app/src/main/res/layout/activity_web.xml
@@ -6,5 +6,5 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activity.WebActivity">
+        tools:context=".activity.WebActivity">
 </WebView>
